refactor: remove HypConv3DHyperboloid and wrap beta_scale in float (#35)
- Remove HypConv3DHyperboloid class (single 3D conv variant was inconsistent with 2D-only coverage)
- Simplify codebase by removing 8 test functions for 3D layer
- Fix HypConv2DPoincare.beta_scale to be a plain Python float instead of bare JAX scalar, resolving nnx.split/merge compatibility issues
- Update documentation and dimension keys to reflect 2D-only convolutions
- Update layer imports and __all__ exports

+++ hyperbolix/nn_layers/poincare_conv.py
@@ -150,7 +150,7 @@ class HypConv2DPoincare(nnx.Module):
         concat_dim = K2 * in_channels
         beta_n = jax.scipy.special.beta(concat_dim / 2.0, 0.5)
         beta_ni = jax.scipy.special.beta(in_channels / 2.0, 0.5)
-        self.beta_scale = beta_n / beta_ni
+        self.beta_scale = float(beta_n / beta_ni)
 
         self.clamping_factor = clamping_factor
         self.smoothing_factor = smoothing_factor
